% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate.R
\name{automate}
\alias{automate}
\alias{automate_make}
\alias{automate_publish}
\alias{automate_docker}
\title{Automate the use of Docker & Make}
\usage{
automate(path = ".")

automate_make(path = ".")

automate_publish(path = ".")

automate_docker(path = ".")
}
\arguments{
\item{path}{Where should we look for RMarkdowns?}
}
\description{
\code{automate()} & friends use yaml metadata from RMarkdowns to create
\code{Dockerfile}'s and \code{Makefile}'s. It should be clear which is created by
\code{automate_docker()} & which by \code{automate_make()}.
}
\seealso{
\code{\link[=automate_load_packages]{automate_load_packages()}}, \code{\link[=automate_load_data]{automate_load_data()}}, \code{\link[=automate_load_scripts]{automate_load_scripts()}}
}
