% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{check_docker}
\alias{check_make}
\alias{check_git}
\alias{check_brew}
\alias{check_choco}
\alias{check_ssh}
\alias{check_github_token}
\alias{check_github_ssh}
\alias{check_github}
\title{Check System Dependencies}
\usage{
check_docker()

check_make()

check_git()

check_brew()

check_choco()

check_ssh(install = getOption("repro.install"))

check_github_token(install = getOption("repro.install"))

check_github_ssh()

check_github()
}
\arguments{
\item{install}{Should something be installed? Defaults to "ask", but can be TRUE/FALSE.}
}
\description{
Check if a dependency is installed and if not it recommends how to install it
depending on the operating system. Most importantly it checks for
\code{git}, \code{make} & \code{docker}. And just for convinience of the installation it
checks on OS X for \code{Homebrew} and on Windows for \code{Chocolately}.
}
\seealso{
Other checkers: 
\code{\link{has}}
}
\concept{checkers}
