% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brew.R
\name{check_brew}
\alias{check_brew}
\title{Check Homewbrew}
\usage{
check_brew(install = TRUE)
}
\arguments{
\item{install}{Should we give recommendations on the installation?}
}
\description{
Check if Homewbrew is installed and if not it recommends how to install it.
}
