% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_choco.R
\name{check_choco}
\alias{check_choco}
\title{Check Homewbrew}
\usage{
check_choco(install = TRUE)
}
\arguments{
\item{install}{Should we give recommendations on the installation?}
}
\description{
Check if Homewbrew is installed and if not it recommends how to install it.
}
