% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docker.R
\name{check_docker}
\alias{check_docker}
\title{Check Docker}
\usage{
check_docker(install = TRUE)
}
\arguments{
\item{install}{Should we give recommendations on the installation?}
}
\description{
Check if docker is installed and if not it recommends how to install it
depending on the operating system.
}
