% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_package.R
\name{check_package}
\alias{check_package}
\title{Check if package exists}
\usage{
check_package(pkg, install = getOption("repro.install"), github = NULL)
}
\arguments{
\item{pkg}{Which package are we talking about?}

\item{install}{Should we install the package in case its missing?}

\item{github}{A github username/package from which the package is installed.
If NULL (the default) CRAN (or whatever repo) you have set is used.}
}
\description{
Check if package exists
}
\seealso{
Other checkers: 
\code{\link{check}},
\code{\link{has_package}()},
\code{\link{has}},
\code{\link{uses}}
}
\concept{checkers}
