% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{current_hash}
\alias{current_hash}
\title{Report Current Hash}
\usage{
current_hash(length = 7L, backend = NULL)
}
\arguments{
\item{length}{The length to which the hash is cut.}

\item{backend}{Can be either of \code{NULL}, "gert", or "git". If \code{NULL} Git is
used with preference and \code{gert} as a fallback.}
}
\description{
Find out what the currently checked out commit is and report its respective
hash.
}
