% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R, R/has_package.R
\name{has}
\alias{has}
\alias{has_make}
\alias{has_git}
\alias{has_docker}
\alias{has_docker_running}
\alias{has_choco}
\alias{has_brew}
\alias{has_ssh}
\alias{has_github_token}
\alias{has_github_token_access}
\alias{has_github_ssh}
\alias{has_github}
\alias{has_renv}
\alias{has_targets}
\alias{has_worcs}
\alias{has_gert}
\title{Check System Dependencies}
\usage{
has_make(silent = TRUE, force_logical = TRUE)

has_git(silent = TRUE, force_logical = TRUE)

has_docker(silent = TRUE, force_logical = TRUE)

has_docker_running(silent = TRUE, force_logical = TRUE)

has_choco(silent = TRUE, force_logical = TRUE)

has_brew(silent = TRUE, force_logical = TRUE)

has_ssh(silent = TRUE, force_logical = TRUE)

has_github_token(silent = TRUE, force_logical = TRUE)

has_github_token_access(silent = TRUE, force_logical = TRUE)

has_github_ssh(silent = TRUE, force_logical = TRUE)

has_github(silent = TRUE, force_logical = TRUE)

has_renv(silent = TRUE, force_logical = TRUE)

has_targets(silent = TRUE, force_logical = TRUE)

has_worcs(silent = TRUE, force_logical = TRUE)

has_gert(silent = TRUE, force_logical = TRUE)
}
\arguments{
\item{silent}{Should a message be printed that informs the user?}

\item{force_logical}{Should the return value be passed through \href{base}{\code{isTRUE}}?}
}
\description{
Corresponding functions to \code{\link{check}}, but intended for programatic use.
}
\seealso{
Other checkers: 
\code{\link{check_package}()},
\code{\link{check}},
\code{\link{has_package}()},
\code{\link{uses}}
}
\concept{checkers}
