% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_package.R
\name{has_package}
\alias{has_package}
\title{Check if package exists}
\usage{
has_package(pkg, silent = TRUE, force_logical = TRUE)
}
\arguments{
\item{pkg}{Which package are we talking about?}

\item{silent}{Should a message be printed that informs the user?}

\item{force_logical}{Should the return value be passed through \href{base}{\code{isTRUE}}?}
}
\description{
Check if package exists
}
\seealso{
Other checkers: 
\code{\link{check_package}()},
\code{\link{check}},
\code{\link{has}}
}
\concept{checkers}
