% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{use_make}
\alias{use_make_docker}
\alias{use_make_singularity}
\title{Use Make}
\usage{
use_make(docker = FALSE, singularity = FALSE, torque = FALSE, open = TRUE)

use_make_docker(use_docker = TRUE, dockerignore = TRUE, open = FALSE)

use_make_singularity(use_singularity = TRUE, open = FALSE)
}
\arguments{
\item{docker}{If true a setup is created that can partially send make commands to a Docker container}

\item{singularity}{If true a setup is created that can partially send make commands to a Singularity container (which requires the Dockerimage)}

\item{torque}{If true a setup is created that can partially send make comands to a TORQUE job scheduler. Especially usefull in combination with a Singularity container.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}

\item{use_docker}{If true \code{use_docker()} is called.}

\item{dockerignore}{If true a .dockerignore file is created.}

\item{use_singularity}{If true \code{use_singularity()} is called.}
}
\description{
Add a (GNU-)Makefile(s) with special emphasis on the use of containers.
}
