% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{repro_template}
\alias{repro_template}
\title{Create repro template}
\usage{
repro_template(path, ...)
}
\arguments{
\item{path}{character specifieng target folder. Will be created if it doesn't exist.}

\item{...}{Passed down to the markdown template.}
}
\description{
Fills a newly created folder with examples of RMarkdown, scripts, and data, which work well with \code{\link[=automate]{automate()}}.
}
