% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\name{reproduce}
\alias{reproduce}
\alias{reproduce_make}
\title{Find entrypoint for analysis and suggest how to reproduce it.}
\usage{
reproduce(fun, ..., path = ".", cache = FALSE, silent = FALSE)

reproduce_make(path, cache = FALSE, silent = FALSE)
}
\arguments{
\item{fun}{a function that inspects \code{dir} and advises on how to reproduce the analysis. Defaults to \link{reproduce_funs}.}

\item{...}{more functions like \code{fun}.}

\item{path}{Were should I look for entrypoints?}

\item{cache}{Default is \code{FALSE}. Some entrypoints have a cache, which you probably do not want to use in a reproduction.}

\item{silent}{Should a message be presented?}
}
\value{
Returns invisibly \code{TRUE} if an entrypoint was found and \code{FALSE} if not.
}
\description{
\code{reproduce()} inspects the files of a project and suggest a way to reproduce the project.
}
\details{
\code{reproduce()} walks through a list of functions that check for a specific entrypoint.
As soon as a function returns a possible entrypoint the search stops.
If no function is supplied the standard list of \link{reproduce_funs} is used.
}
\seealso{
reproduce_funs
}
