% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce.R
\docType{data}
\name{reproduce_funs}
\alias{reproduce_funs}
\title{A list of functions to detect entrypoints.}
\format{
An object of class \code{list} of length 1.
}
\usage{
reproduce_funs
}
\description{
At the moment only reproduce_make is available.
}
\details{
\code{reproduce_make} detects make as an entrypoint if there is a Makefile at top level.
If it does also encounter a Makefile_Docker somewhere it recognizes the different make instructions.
}
\seealso{
reproduce
}
\keyword{datasets}
