% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{rerun}
\alias{rerun}
\title{Find entrypoint for analysis and suggest how to reproduce it.}
\usage{
rerun(...)
}
\arguments{
\item{...}{passed to \code{\link[=reproduce]{reproduce()}}.}
}
\description{
\code{rerun()} was renamed to \code{\link[=reproduce]{reproduce()}}. \code{rerun()} is deprecated.
}
\seealso{
reproduce_funs
}
