% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\docType{data}
\name{rerun_funs}
\alias{rerun_funs}
\title{A list of functions to detect entrypoints.}
\format{
An object of class \code{list} of length 1.
}
\usage{
rerun_funs
}
\description{
At the moment only rerun_make is available.
}
\details{
\code{rerun_make} detects make as an entrypoint if there is a Makefile at top level.
If it does also encounter a Makefile_Docker somewhere it recognizes the different make instructions.
}
\seealso{
rerun
}
\keyword{datasets}
