% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_gha_docker}
\alias{use_gha_docker}
\title{Use GitHub Action to Build Dockerimage}
\usage{
use_gha_docker(file = getOption("repro.gha.docker"), open = TRUE)
}
\arguments{
\item{file}{Which file to save to.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}
}
\description{
Add an standard actions that builds and publishes an Dockerimage from the \code{Dockerfile}.
}
