% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_gha_publish}
\alias{use_gha_publish}
\title{Use GitHub Action to Publish Results}
\usage{
use_gha_publish(file = getOption("repro.gha.publish"), open = TRUE)
}
\arguments{
\item{file}{Which file to save to.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}
}
\description{
Add an standard actions that builds the \code{publish}-target, and publishes the results on the \code{gh-pages} branch.
Requires a Dockerimage published within the same GitHub repository. See \code{\link[=use_gha_docker]{use_gha_docker()}}.
}
