% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_github_action_docker}
\alias{use_github_action_docker}
\title{Use GitHub Action to Build Dockerimage}
\usage{
use_github_action_docker(
  file = ".github/workflows/push-container.yml",
  open = TRUE
)
}
\arguments{
\item{file}{Which file to save to.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}
}
\description{
Add an standard actions that builds and publishes an Dockerimage from the \code{Dockerfile}.
}
