% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gha.R
\name{use_template_template}
\alias{use_template_template}
\title{Deal with templates that are themselves whisker templates.}
\usage{
use_template_template(
  template,
  save_as = template,
  escape = c(`\\\\[\\\\[` = "{{", `\\\\]\\\\]` = "}}"),
  data = list(),
  ignore = FALSE,
  open = FALSE,
  package = "repro"
)
}
\arguments{
\item{template}{\code{\link[usethis:use_template]{usethis::use_template()}}}

\item{save_as}{\code{\link[usethis:use_template]{usethis::use_template()}}}

\item{escape}{named vector, name is replacement (default curly), value is placeholder (default square)}

\item{data}{\code{\link[usethis:use_template]{usethis::use_template()}}}

\item{ignore}{\code{\link[usethis:use_template]{usethis::use_template()}}}

\item{open}{\code{\link[usethis:use_template]{usethis::use_template()}}}

\item{package}{\code{\link[usethis:use_template]{usethis::use_template()}}}
}
\description{
If a template is itself used again as a whisker template, we need some other escape mechanism.
The \code{repro}-template than uses \verb{[[]]} instead of \code{{{}}} which are than later reinstated.
}
