% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uses.R
\name{uses}
\alias{uses}
\alias{uses_make}
\alias{uses_docker}
\title{Check Project Configuration}
\usage{
uses_make(silent = FALSE)

uses_docker(silent = FALSE)
}
\arguments{
\item{silent}{Defaults to \code{false}. Should a message be printed that informs the user?}
}
\description{
Check if a project uses certain things, e.g. Make or Docker.
}
\seealso{
Other checkers: 
\code{\link{check_package}()},
\code{\link{check}},
\code{\link{has_package}()},
\code{\link{has}}
}
\concept{checkers}
