% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate.R
\name{automate_load}
\alias{automate_load}
\alias{automate_load_packages}
\alias{automate_load_scripts}
\alias{automate_load_data}
\title{Access repro YAML Metadata from within the document}
\usage{
automate_load_packages()

automate_load_scripts()

automate_load_data(data, func, ...)
}
\arguments{
\item{data}{How is the entry in the YAML called? It will be the name of the object.}

\item{func}{Which function should be used to read in the data? Its first argument must be the path to the file.}

\item{...}{Further arguments supplied to \code{func}.}
}
\value{
\code{automate_load_packages()} & \code{automate_load_scripts()} do not return anything. \code{automate_load_data()} returns the data.
}
\description{
\itemize{
\item \code{automate_load_packages()} loads all packages listed in YAML via \code{library()}
\item \code{automate_load_scripts()} registeres external scripts via \code{knitr::read_chunk()}
\item \code{automate_load_data()} reads in the data from the yaml with abitrary functions
}
}
