% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docker.R
\name{check}
\alias{check}
\alias{check_docker}
\alias{check_brew}
\alias{check_choco}
\title{Check System Dependencies}
\usage{
check_docker(install = TRUE)

check_brew(install = TRUE)

check_choco(install = TRUE)
}
\arguments{
\item{install}{Should we give recommendations on the installation?}
}
\description{
Check if a dependency is installed and if not it recommends how to install it
depending on the operating system. Most importantly it checks for
\code{git}, \code{make} & \code{docker}. And just for convinience of the installation it
checks on OS X for \code{Homebrew} and on Windows for \code{Chocolately}.
}
