% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/check_package.R
\name{check}
\alias{check}
\alias{check_docker}
\alias{check_make}
\alias{check_git}
\alias{check_brew}
\alias{check_choco}
\alias{check_ssh}
\alias{check_github_token}
\alias{check_github_token_access}
\alias{check_github_ssh}
\alias{check_github}
\alias{check_renv}
\alias{check_targets}
\alias{check_worcs}
\title{Check System Dependencies}
\usage{
check_docker()

check_make()

check_git()

check_brew()

check_choco()

check_ssh(install = getOption("repro.install"))

check_github_token(install = getOption("repro.install"))

check_github_token_access()

check_github_ssh()

check_github(auth_method = "token")

check_renv(install = getOption("repro.install"))

check_targets(install = getOption("repro.install"))

check_worcs(install = getOption("repro.install"))
}
\arguments{
\item{install}{Should something be installed? Defaults to "ask", but can be TRUE/FALSE.}

\item{auth_method}{How do you want to authenticate with GitHub? Either "token" (default) or "ssh".}
}
\description{
Check if a dependency is installed and if not it recommends how to install it
depending on the operating system. Most importantly it checks for
\code{git}, \code{make} & \code{docker}. And just for convenience of the installation it
checks on OS X for \code{Homebrew} and on Windows for \code{Chocolately}.
}
\seealso{
Other checkers: 
\code{\link{check_package}()},
\code{\link{has_package}()},
\code{\link{has}},
\code{\link{uses}}
}
\concept{checkers}
