% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_windows_path.R
\name{docker_windows_path}
\alias{docker_windows_path}
\title{Genereate the weird format for docker on windows.}
\usage{
docker_windows_path(path = NULL)
}
\arguments{
\item{path}{Some windows path. If NULL the current directory.}
}
\description{
Genereate the weird format for docker on windows.
}
