% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has.R
\name{has}
\alias{has}
\alias{has_make}
\alias{has_git}
\alias{has_docker}
\alias{has_choco}
\alias{has_brew}
\title{Check System Dependencies}
\usage{
has_make(silent = TRUE, allow_na = FALSE)

has_git(silent = TRUE, allow_na = FALSE)

has_docker(silent = TRUE, allow_na = FALSE)

has_choco(silent = TRUE, allow_na = FALSE)

has_brew(silent = TRUE, allow_na = FALSE)
}
\arguments{
\item{silent}{Should a message be printed that informs the user?}

\item{allow_na}{Is \code{NA} a valid return value? If \code{NA} is not allowed, FALSE is returned instead.}
}
\description{
Corresponding functions to \code{\link{check}}, but intended for programatic use.
}
\seealso{
Other checkers: 
\code{\link{check}}
}
\concept{checkers}
