% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{use_make}
\alias{use_make_docker}
\alias{use_make_singularity}
\alias{use_make_publish}
\title{Use Make}
\usage{
use_make(
  docker = FALSE,
  publish = FALSE,
  singularity = FALSE,
  torque = FALSE,
  open = TRUE
)

use_make_docker(file, use_docker = TRUE, dockerignore = TRUE, open = FALSE)

use_make_singularity(file, use_singularity = TRUE, open = FALSE)

use_make_publish(file, open = FALSE)
}
\arguments{
\item{docker}{If true or a path a setup is created that can partially send make commands to a Docker container.}

\item{publish}{Should the \code{Makefile_publish} also be created?}

\item{singularity}{If true or a path a setup is created that can partially send make commands to a Singularity container (which requires the Dockerimage)}

\item{torque}{If true a or a path setup is created that can partially send make comands to a TORQUE job scheduler. Especially usefull in combination with a Singularity container.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}

\item{file}{Path to the file that is to be created.}

\item{use_docker}{If true \code{use_docker()} is called.}

\item{dockerignore}{If true a .dockerignore file is created.}

\item{use_singularity}{If true \code{use_singularity()} is called.}
}
\description{
Add a (GNU-)Makefile(s) with special emphasis on the use of containers.
}
