% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{rerun}
\alias{rerun}
\alias{rerun_make}
\title{Find entrypoint for analysis and suggest how to rerun it.}
\usage{
rerun(fun, ..., path = ".", cache = FALSE, silent = FALSE)

rerun_make(path, cache = FALSE, silent = FALSE)
}
\arguments{
\item{fun}{a function that inspects \code{dir} and advises on how to rerun the analysis. Defaults to \link{rerun_funs}.}

\item{...}{more functions like \code{fun}.}

\item{path}{Were should I look for entrypoints?}

\item{cache}{Default is \code{FALSE}. Some entrypoints have a cache, which you probably do not want to use in a reproduction.}

\item{silent}{Should a message be presented?}
}
\value{
Returns invisibly \code{TRUE} if an entrypoint was found and \code{FALSE} if not.
}
\description{
\code{rerun()} inspects the files of a project and suggest a way to reproduce the project.
}
\details{
\code{rerun()} walks through a list of functions that check for a specific entrypoint.
As soon as a function returns a possible entrypoint the search stops.
If no function is supplied the standard list of \link{rerun_funs} is used.
}
\seealso{
rerun_funs
}
