% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{use_docker}
\alias{use_docker}
\title{Use Docker}
\usage{
use_docker(
  rver = NULL,
  stack = "verse",
  date = Sys.Date(),
  file = "Dockerfile",
  open = TRUE
)
}
\arguments{
\item{rver}{Which r version to use, defaults to current version.}

\item{stack}{Which stack to use, possible values are \code{c("r-ver", "rstudio", "tidyverse", "verse", "geospatial")}.}

\item{date}{Which date should be used for package instalation, defaults to today.}

\item{file}{Which file to save to}

\item{open}{Open the newly created file for editing? Happens in RStudio, if applicable, or via utils::file.edit() otherwise.}
}
\description{
Add or modify the Dockerfile in the current project.
}
