% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{use_docker}
\alias{use_docker}
\title{Use Docker}
\usage{
use_docker(rver = NULL, stack = "verse", date = Sys.Date())
}
\arguments{
\item{rver}{which r version to use, defaults to current version.}

\item{stack}{which stack to use, possible values are \code{c("r-ver", "rstudio", "tidyverse", "verse", "geospatial")}.}

\item{date}{which date should be used for package instalation, defaults to today.}
}
\description{
Add or modify the Dockerfile in the current project.
}
