% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{use_docker_packages}
\alias{use_docker_packages}
\title{Add dependencies to Dockerfile}
\usage{
use_docker_packages(
  packages,
  github = NULL,
  strict = TRUE,
  write = TRUE,
  open = write
)
}
\arguments{
\item{packages}{Which packages to add.}

\item{github}{Are there github packages?}

\item{strict}{Defaults to TRUE, force a specific version for github packages.}

\item{write}{Should the file be written to disk?}

\item{open}{Should the file be opened?}
}
\description{
Adds package dependencies as a new RUN statement to Dockerfile.
Sorts packages first into source (cran & github) and then alphabetically.
}
