% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{use_docker_packages}
\alias{use_docker_packages}
\title{Add dependencies to Dockerfile}
\usage{
use_docker_packages(packages, github = NULL, strict = TRUE,
  open = TRUE)
}
\arguments{
\item{packages}{Which packages to add.}

\item{github}{Are there github packages?}

\item{strict}{Defaults to TRUE, force a specific version for github packages.}

\item{open}{Should the \code{Dockerfile} be opened?}
}
\description{
Adds package dependencies as a new RUN statement to Dockerfile.
Sorts packages first into source (cran & github) and then alphabetically.
}
